+++ product/src/test/java/com/atticket/product/service/ProductServiceTest.java
@@ -4,8 +4,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.Optional;
 
 import org.junit.jupiter.api.Assertions;
@@ -16,10 +14,6 @@ import org.junit.jupiter.api.Test;
 import com.atticket.common.response.BaseException;
 import com.atticket.product.domain.Product;
 import com.atticket.product.repository.ProductRepository;
-import com.atticket.product.type.AgeLimit;
-import com.atticket.product.type.Category;
-import com.atticket.product.type.Region;
-import com.atticket.product.type.SubCategory;
 
 public class ProductServiceTest {
 
@@ -37,29 +31,14 @@ public class ProductServiceTest {
 
 		//Given
 		Long productId = 1L;
+		Product givenProduct = Product.builder().build();
 
-		Product givenProduct = Product.builder()
-			.id(productId)
-			.category(Category.MUSICAL)
-			.subCategory(SubCategory.ORIGINAL)
-			.name("상품1")
-			.explain("설명")
-			.runningTime(LocalTime.of(2, 0))
-			.startDate(LocalDate.of(2023, 4, 21))
-			.endDate(LocalDate.of(2024, 4, 21))
-			.ageLimit(AgeLimit.FIFTEEN)
-			.image("http://이미지.jpg")
-			.region(Region.SEOUL)
-			.build();
-
+		//When
 		when(productRepository.findById(productId)).thenReturn(Optional.of(givenProduct));
-
-		//when
 		Product result = productService.getProductById(productId);
 
-		Assertions.assertEquals(result.getId(), givenProduct.getId());
-		Assertions.assertEquals(result.getName(), givenProduct.getName());
-
+		//Then
+		Assertions.assertEquals(result, givenProduct);
 		verify(productRepository).findById(1L);
 
 	}
@@ -68,29 +47,31 @@ public class ProductServiceTest {
 	@DisplayName("상품 삭제")
 	void deleteProduct() {
 
+		//Given
 		Long productId = 1L;
+		Product givenProduct = Product.builder().build();
 
-		Product givenProduct = Product.builder()
-			.id(productId)
-			.category(Category.MUSICAL)
-			.subCategory(SubCategory.ORIGINAL)
-			.name("상품1")
-			.explain("설명")
-			.runningTime(LocalTime.of(2, 0))
-			.startDate(LocalDate.of(2023, 4, 21))
-			.endDate(LocalDate.of(2024, 4, 21))
-			.ageLimit(AgeLimit.FIFTEEN)
-			.image("http://이미지.jpg")
-			.region(Region.SEOUL)
-			.build();
-
+		//When
 		when(productRepository.findById(productId)).thenReturn(Optional.of(givenProduct));
-
 		productService.deleteProduct(productId);
 
+		//Then
+		verify(productRepository).deleteById(productId);
+	}
+
+	@Test
+	@DisplayName("상품 삭제 에러")
+	void deleteProductNotFindId() {
+
+		//Given
+		Long productId = 1L;
+
+		//When
 		when(productRepository.findById(productId)).thenReturn(null);
 
+		//Then
 		Assertions.assertThrows(BaseException.class, () -> productService.deleteProduct(productId));
 
 	}
+
 }
