+++ reservation/src/main/resources/application-dev.yml
@@ -1,2 +1,2 @@
 server:
-  port: 9001
+  port: 8101
