+++ reservation/src/main/resources/application-local.yml
@@ -1,2 +1,2 @@
 server:
-  port: 9000
+  port: 8100
