+++ reservation/src/main/resources/application-prod.yml
@@ -1,2 +1,2 @@
 server:
-  port: 9002
+  port: 8102
